Change int level type to size_t

+++ include/LevelAccess.hh
@@ -26,7 +26,7 @@ private:
 	inline static void set_levels(MultilevelConfig& config, std::vector<LevelDef*> levels) {
 		config.set_levels(levels);
 	}
-	inline static void update(MultilevelConfig& config, int level) {
+	inline static void update(MultilevelConfig& config, size_t level) {
 		config.update(level);
 	}
